requirements.txt
# requirements.txt

# Web Framework
fastapi
uvicorn[standard]

# LLM and Vector DB Clients
google-generativeai
pinecone-client
langchain-text-splitters

# Document Processing
requests
pymupdf
python-docx

# Configuration
python-dotenv

# Database
sqlalchemy
psycopg2-binary
alembic
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv